+++ misc/ch03/BadIndent.hs
@@ -1,0 +1,8 @@
+-- file: Ch03/BadIndent.hs
+-- This is the leftmost column.
+
+    -- Our first declaration is in in column 4
+    firstBadIndentation = 1
+
+  -- Our second is left of the first, which is illegal!!
+  secondBadIndentation = 2