+++ src/Ch03/GlobalVariable.hs
@@ -1,1 +1,16 @@
+-- |
+-- Module      :  Ch03.GlobalVariable
+-- Description :  Local Functions, Global Variables
+-- Copyright   :  erlnow 2020 - 2030
+-- License     :  BSD3
+--
+-- Maintainer  :  erlestau@gmail.com
+-- Stability   :  experimental
+-- Portability :  unknown
+--
+-- Example from Chapter 3: Defining Types, Streamlining functions
+
+module Ch03.GlobalVariable where
+
+
 itemName = "Weighted Companion Cube"