+++ src/Ch03/GoodIndent.hs
@@ -1,0 +1,22 @@
+-- |
+-- Module      :  Ch03.GoodIndent
+-- Description :  The Offside Rule and Whitespace in an Expression
+-- Copyright   :  erlnow 2020 - 2030
+-- License     :  BSD3
+--
+-- Maintainer  :  erlestau@gmail.com
+-- Stability   :  experimental
+-- Portability :  unknown
+--
+-- Chapter 3: Defining Types, Streamlining Functions
+
+module Ch03.GoodIndent where
+
+-- p. 64
+
+
+  -- It's fine for top-level declarations to start in any column...
+  firstGoodIndetantion = 1
+
+  -- ...provided all subsequents declarations do, too!
+  secondGoodIndentation = 2